Combined tests for ec2_availability and ec2_limits
